Convert service-broker to use CRD for creating messaginguser (#3649)
* Convert service-broker to use CRD for creating messaginguser

This allows removing a lot of code for interacting directly with
Keycloak in keycloak-user-api
